requirements.txt
requests==2.26.0
python-telegram-bot==13.13